Make it possible to not complete a request when the event source is done.
Abstract the event source of a stream queue so it doesn't have to be a stream.

This is in anticipation of adding look-ahead which will need to keep the
request alive until the user fast-forwards or rewindes, even after the
original event source is done, and which will need a stream queue fed by
the look-ahead request, not by a stream.

R=nweiz@google.com

Review URL: https://codereview.chromium.org//1305063002 .