Make `TypeSafeStream` extend `Stream`.
This makes it automatically compatible with future additions to `Stream`,
like the `groupBy` method which will be added in SDK version 1.24.
It inherits the default implementation, which will use the `listen` method
of the `TypeSafeStream`, returning a `TypeSafeStreamSubscription`, and will
just have a type check for each event before the listener even sees it.

BUG=
R=floitsch@google.com

Review-Url: https://codereview.chromium.org//2872233002 .