Fix an ArgumentError constructor call
The arguments that were passed match the intention of the
`ArgumentError` constructor, while the `ArgumentError.value` constructor
reverses the message and argument name, and includes the argument value.
Including the value is not useful in this case since it is always `[]`
and in most cases won't be the exact instance that was passed because of
the `.toList()` call which would usually construct a copy.

Bump the min SDK constraint to `2.14` which is the first release
allowing the `name` argument to the `ArgumentError` constructor.