Deprecate `AsyncCache.fetchStream`.
The feature complicates the `AsyncCache` class, and would be a better fit
for a separate class.
The functionality is not great. Since it returns the "same" stream
each time the cache is queried (until evaluated), it actually
caches all stream events
The invalidation is tricky because the timer doesn't start until
you *listen* to the stream. In practice, people will usually
listen to the stream they fetch.

It also seems that no-one ever used the functionality.
There is no use inside Google, or in [Github](https://github.com/search?q=%22.fetchStream%22+language%3ADart+-filename%3Aasync_cache.dart+-filename%3Aasync_cache_test.dart&type=code),
outside of the package's own tests.

It's a better fit for a separate class, but it also appears to
not be used anywhere