Simplify implementation of `StreamQueue`.
Only change in behavior is that all event seen by the original queue
while the transaction request is active, are forwarded to the transaction
immediately, and will still be visible after the transaction commits
or rejects.
The existing behavior was based on cancelling a stream subscription
and which events were visible depended on which commands had been
used to request those events.