Add `StreamExtensions.listenAndBuffer()` (#235)
The immediate motivation for this is to buffer stderr from a subprocess until we learn whether the subprocess completed successfully or not. It's important to consume the output eagerly so the process doesn't deadlock after saturating the OS and/or `dart:io` buffers, but we don't want to print it at all if the process completes successfully.